pyproject.toml
[tool.poetry]
name = "langgraph-article-writer-critiquer"
version = "0.1.0"
description = "Example of using Langgraph to write and critique articles"
authors = [
    "langchain-ai"
]
packages = [
    { include = "langgraph-article-writer-critiquer" },
]

[tool.poetry.dependencies]
python = ">=3.9.0,<3.13"
langgraph = "^0.2.0"
langchain_anthropic = "^0.1.0"
langchain_core = "^0.2.0"
langchain_openai = "^0.1.0"
tavily-python = "^0.3.0"
langchain_community = "^0.2.0"
langchain_ollama = "^0.1.3"
python-dotenv = "^1.0.1"
ruff = "^0.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"